Give more flexibility to some damage configs (sword, pickaxe, hoe) than vanilla allows as we override a bunch of things anyways so it doesn't matter it only supports ints in those places #6968. Also open other damage configs up to going negative to the point of the base tier's damage value to allow things to mimic the hoe as zero damage. One thing to note is the config values for these options will likely get "corrected" to the defaults so they will need to be changed again as desired.

+++ src/tools/java/mekanism/tools/common/item/ItemMekanismHoe.java
@@ -34,7 +34,7 @@ public class ItemMekanismHoe extends HoeItem implements IHasRepairType, IAttribu
     private final AttributeCache attributeCache;
 
     public ItemMekanismHoe(MaterialCreator material, Item.Properties properties) {
-        super(material, material.getHoeDamage(), material.getHoeAtkSpeed(), properties);
+        super(material, (int) material.getHoeDamage(), material.getHoeAtkSpeed(), properties);
         this.material = material;
         this.attributeCache = new AttributeCache(this, material.attackDamage, material.hoeDamage, material.hoeAtkSpeed);
     }
